requirements.txt
black==20.8b1
Jinja2==2.11.3
Pillow==8.2.0
